refactor(test): app의 중복 ServiceTest·ServiceTestConfig를 test-support로 통합
